refactor: use existing usual integration
