--- src/dex/m-token/m-token.ts
@@ -1,167 +1,0 @@
-import { AsyncOrSync } from 'ts-essentials';
-import type {
-  Token,
-  Address,
-  ExchangePrices,
-  PoolPrices,
-  AdapterExchangeParam,
-  Logger,
-  PoolLiquidity,
-} from '../../types';
-import { SwapSide, Network } from '../../constants';
-import * as CALLDATA_GAS_COST from '../../calldata-gas-cost';
-import { IDex } from '../idex';
-import { IDexHelper } from '../../dex-helper/idex-helper';
-import type { MTokenData, DexParams } from './types';
-import { SimpleExchange } from '../simple-exchange';
-import { BI_POWS } from '../../bigint-constants';
-
-export class MToken extends SimpleExchange implements IDex<MTokenData> {
-  readonly hasConstantPriceLargeAmounts = true;
-
-  readonly needWrapNative = false;
-
-  readonly isFeeOnTransferSupported = false;
-
-  logger: Logger;
-
-  constructor(
-    readonly network: Network,
-    readonly dexKey: string,
-    readonly dexHelper: IDexHelper,
-    readonly config: DexParams,
-  ) {
-    super(dexHelper, dexKey);
-    this.logger = dexHelper.getLogger(dexKey);
-  }
-
-  // No initialization needed for constant price
-  async initializePricing() {}
-
-  // Legacy: was only used for V5
-  // Returns the list of contract adapters (name and index)
-  // for a buy/sell. Return null if there are no adapters.
-  getAdapters() {
-    return null;
-  }
-
-  // Returns list of pool identifiers that can be used
-  // for a given swap. poolIdentifiers must be unique
-  // across DEXes. It is recommended to use
-  // ${dexKey}_${poolAddress} as a poolIdentifier
-  async getPoolIdentifiers(
-    from: Token,
-    to: Token,
-    side: SwapSide,
-    blockNumber: number,
-  ): Promise<string[]> {
-    if (!this.ensureOrigin({ from, to })) {
-      return [];
-    }
-
-    return [`${this.dexKey}_${this.config.WrappedM.address}`];
-  }
-
-  // Returns pool prices for amounts.
-  // If limitPools is defined only pools in limitPools
-  // should be used. If limitPools is undefined then
-  // any pools can be used.
-  async getPricesVolume(
-    from: Token,
-    to: Token,
-    amounts: bigint[],
-    side: SwapSide,
-    blockNumber: number,
-    limitPools?: string[],
-  ): Promise<null | ExchangePrices<MTokenData>> {
-    if (!this.ensureOrigin({ from, to })) {
-      return null;
-    }
-
-    // 1:1 swap
-    // Amounts need no adjustment
-    const unitOut = BI_POWS[to.decimals];
-
-    return [
-      {
-        unit: unitOut,
-        prices: amounts,
-        data: {},
-        poolAddresses: [this.config.WrappedM.address],
-        exchange: this.dexKey,
-        gasCost: 70000,
-        poolIdentifier: this.dexKey,
-      },
-    ];
-  }
-
-  // Returns estimated gas cost of calldata for this DEX in multiSwap
-  getCalldataGasCost(poolPrices: PoolPrices<MTokenData>): number | number[] {
-    return CALLDATA_GAS_COST.DEX_NO_PAYLOAD;
-  }
-
-  // Encode params required by the exchange adapter
-  // V5: Used for multiSwap, buy & megaSwap
-  // V6: Not used, can be left blank
-  // Hint: abiCoder.encodeParameter() could be useful
-  getAdapterParam(
-    srcToken: string,
-    destToken: string,
-    srcAmount: string,
-    destAmount: string,
-    data: MTokenData,
-    side: SwapSide,
-  ): AdapterExchangeParam {
-    const exchange = this.config.WrappedM.address;
-    const payload = '0x';
-
-    return {
-      targetExchange: exchange,
-      payload,
-      networkFee: '0',
-    };
-  }
-
-  // Returns list of top pools based on liquidity. Max
-  // limit number pools should be returned.
-  async getTopPoolsForToken(
-    tokenAddress: Address,
-    limit: number,
-  ): Promise<PoolLiquidity[]> {
-    if (!this.isMOrWrappedM(tokenAddress)) {
-      return [];
-    }
-
-    const isM =
-      tokenAddress.toLowerCase() === this.config.MToken.address.toLowerCase();
-
-    return [
-      {
-        exchange: this.dexKey,
-        address: this.config.WrappedM.address,
-        connectorTokens: [isM ? this.config.WrappedM : this.config.MToken],
-        liquidityUSD: 1000000000, // Returning a big number to prefer this DEX
-      },
-    ];
-  }
-
-  // Checks if the token addresses is M or WrappedM
-  isMOrWrappedM(tokenAddress: Address): boolean {
-    return (
-      tokenAddress.toLowerCase() === this.config.MToken.address.toLowerCase() ||
-      tokenAddress.toLowerCase() === this.config.WrappedM.address.toLowerCase()
-    );
-  }
-
-  // Ensures that the token addresses are correct
-  ensureOrigin(args: Partial<{ from: Token; to: Token }>): boolean {
-    const areDifferentAddresses =
-      args?.from?.address?.toLowerCase() !== args?.to?.address?.toLowerCase();
-
-    return (
-      areDifferentAddresses &&
-      this.isMOrWrappedM(args?.from?.address || '') &&
-      this.isMOrWrappedM(args?.to?.address || '')
-    );
-  }
-}
