--- src/dex/m-token/m-wrapped-m.ts
@@ -1,68 +1,0 @@
-import { Network, SwapSide } from '../../constants';
-import { getDexKeysWithNetwork } from '../../utils';
-import { IDexHelper } from '../../dex-helper/idex-helper';
-import { MToken } from './m-token';
-import type {
-  Address,
-  DexExchangeParam,
-  NumberAsString,
-  Token,
-} from '../../types';
-import { Interface } from '@ethersproject/abi';
-import WRAPPED_M_ABI from '../../abi/m-token/WrappedM.abi.json';
-import { Config } from './config';
-
-export class MWrappedM extends MToken {
-  public static dexKeysWithNetwork: { key: string; networks: Network[] }[] =
-    getDexKeysWithNetwork(Config);
-
-  wrappedMInterface: Interface;
-
-  constructor(
-    readonly network: Network,
-    readonly dexKey: string,
-    readonly dexHelper: IDexHelper,
-  ) {
-    const config = Config[dexKey][network];
-    super(network, dexKey, dexHelper, config);
-
-    this.wrappedMInterface = new Interface(WRAPPED_M_ABI);
-  }
-
-  async getDexParam(
-    from: Address,
-    to: Address,
-    srcAmount: NumberAsString,
-    destAmount: NumberAsString,
-    recipient: Address,
-    data: {},
-    side: SwapSide,
-  ): Promise<DexExchangeParam> {
-    if (
-      !this.ensureOrigin({
-        from: { address: from } as Token,
-        to: { address: to } as Token,
-      })
-    ) {
-      throw new Error('Unexpected token addresses');
-    }
-
-    const fn =
-      from.toLowerCase() === this.config.MToken.address.toLowerCase()
-        ? 'wrap(address, uint256)'
-        : 'unwrap(address, uint256)';
-
-    const exchangeData = this.wrappedMInterface.encodeFunctionData(fn, [
-      recipient,
-      srcAmount,
-    ]);
-
-    return {
-      needWrapNative: false,
-      dexFuncHasRecipient: true,
-      exchangeData,
-      targetExchange: this.config.WrappedM.address,
-      returnAmountPos: undefined,
-    };
-  }
-}
