+++ src/dex/usual/usual-e2e.test.ts
@@ -25,6 +25,7 @@ function testForNetwork(
 
   const sideToContractMethods = new Map([
     [SwapSide.SELL, [ContractMethod.swapExactAmountIn]],
+    [SwapSide.BUY, [ContractMethod.swapExactAmountOut]],
   ]);
 
   describe(`${network}`, () => {
@@ -63,6 +64,52 @@ describe('UsualBond E2E', () => {
 
     const tokenAAmount: string = '100000';
     const tokenBAmount: string = '100000';
+
+    testForNetwork(
+      network,
+      dexKey,
+      tokenASymbol,
+      tokenBSymbol,
+      tokenAAmount,
+      tokenBAmount,
+    );
+  });
+});
+
+describe('WrappedMM E2E', () => {
+  const dexKey = 'WrappedMM';
+
+  describe('Mainnet', () => {
+    const network = Network.MAINNET;
+
+    const tokenASymbol: string = 'WrappedM';
+    const tokenBSymbol: string = 'M';
+
+    const tokenAAmount: string = '1000000000';
+    const tokenBAmount: string = '1000000000';
+
+    testForNetwork(
+      network,
+      dexKey,
+      tokenASymbol,
+      tokenBSymbol,
+      tokenAAmount,
+      tokenBAmount,
+    );
+  });
+});
+
+describe('MWrappedM E2E', () => {
+  const dexKey = 'MWrappedM';
+
+  describe('Mainnet', () => {
+    const network = Network.MAINNET;
+
+    const tokenASymbol: string = 'M';
+    const tokenBSymbol: string = 'WrappedM';
+
+    const tokenAAmount: string = '1000000000';
+    const tokenBAmount: string = '1000000000';
 
     testForNetwork(
       network,
