+++ src/dex/usual/usual-integration.test.ts
@@ -16,6 +16,8 @@ import { UsualMWrappedM } from './usual-m-wrapped-m';
 import { UsualMUsd0 } from './usual-m-usd0';
 import { Usual } from './usual';
 import { UsualBond } from './usual-bond';
+import { WrappedMM } from './wrapped-m-m';
+import { MWrappedM } from './m-wrapped-m';
 
 async function testPricingOnNetwork(
   usual: Usual,
@@ -305,6 +307,200 @@ describe('UsualM<>USD0', function () {
       const usualMUsd0 = new UsualMUsd0(network, dexKey, dexHelper);
 
       const poolLiquidity = await usualMUsd0.getTopPoolsForToken(
+        tokenA.address,
+        10,
+      );
+      console.log(
+        `${tokenA.symbol} Top Pools:`,
+        JSON.stringify(poolLiquidity, null, 2),
+      );
+
+      checkPoolsLiquidity(poolLiquidity, tokenA.address, dexKey);
+    });
+  });
+});
+
+describe('WrappedM<>M', function () {
+  const dexKey = 'WrappedMM';
+  let blockNumber: number;
+  let wrappedMM: WrappedMM;
+
+  describe('Mainnet', () => {
+    const network = Network.MAINNET;
+    const dexHelper = new DummyDexHelper(network);
+
+    // Don't forget to update relevant tokens in constant-e2e.ts
+
+    const amountsForSell = [
+      0n,
+      1n * BI_POWS[18],
+      2n * BI_POWS[18],
+      3n * BI_POWS[18],
+      4n * BI_POWS[18],
+      5n * BI_POWS[18],
+      6n * BI_POWS[18],
+      7n * BI_POWS[18],
+      8n * BI_POWS[18],
+      9n * BI_POWS[18],
+      10n * BI_POWS[18],
+    ];
+
+    beforeAll(async () => {
+      blockNumber = await dexHelper.web3Provider.eth.getBlockNumber();
+      wrappedMM = new WrappedMM(network, dexKey, dexHelper);
+      if (wrappedMM.initializePricing) {
+        await wrappedMM.initializePricing(blockNumber);
+      }
+    });
+
+    it('getPoolIdentifiers and getPricesVolume SELL', async function () {
+      await testPricingOnNetwork(
+        wrappedMM,
+        network,
+        dexKey,
+        blockNumber,
+        'WrappedM',
+        'M',
+        SwapSide.SELL,
+        amountsForSell,
+        '',
+      );
+    });
+
+    it('getPoolIdentifiers and getPricesVolume BUY', async function () {
+      await testPricingOnNetwork(
+        wrappedMM,
+        network,
+        dexKey,
+        blockNumber,
+        'WrappedM',
+        'M',
+        SwapSide.BUY,
+        amountsForSell,
+        '',
+      );
+    });
+
+    it('getTopPoolsForToken: WrappedM', async function () {
+      const tokenA = Tokens[network]['WrappedM'];
+      const dexHelper = new DummyDexHelper(network);
+      const wrappedMM = new WrappedMM(network, dexKey, dexHelper);
+
+      const poolLiquidity = await wrappedMM.getTopPoolsForToken(
+        tokenA.address,
+        10,
+      );
+      console.log(
+        `${tokenA.symbol} Top Pools:`,
+        JSON.stringify(poolLiquidity, null, 2),
+      );
+
+      checkPoolsLiquidity(poolLiquidity, tokenA.address, dexKey);
+    });
+
+    it('getTopPoolsForToken: M', async function () {
+      const tokenA = Tokens[network]['M'];
+      const dexHelper = new DummyDexHelper(network);
+      const wrappedMM = new WrappedMM(network, dexKey, dexHelper);
+
+      const poolLiquidity = await wrappedMM.getTopPoolsForToken(
+        tokenA.address,
+        10,
+      );
+      console.log(
+        `${tokenA.symbol} Top Pools:`,
+        JSON.stringify(poolLiquidity, null, 2),
+      );
+
+      checkPoolsLiquidity(poolLiquidity, tokenA.address, dexKey);
+    });
+  });
+});
+
+describe('M<>WrappedM', function () {
+  const dexKey = 'MWrappedM';
+  let blockNumber: number;
+  let mWrappedM: MWrappedM;
+
+  describe('Mainnet', () => {
+    const network = Network.MAINNET;
+    const dexHelper = new DummyDexHelper(network);
+
+    // Don't forget to update relevant tokens in constant-e2e.ts
+
+    const amountsForSell = [
+      0n,
+      1n * BI_POWS[18],
+      2n * BI_POWS[18],
+      3n * BI_POWS[18],
+      4n * BI_POWS[18],
+      5n * BI_POWS[18],
+      6n * BI_POWS[18],
+      7n * BI_POWS[18],
+      8n * BI_POWS[18],
+      9n * BI_POWS[18],
+      10n * BI_POWS[18],
+    ];
+
+    beforeAll(async () => {
+      blockNumber = await dexHelper.web3Provider.eth.getBlockNumber();
+      mWrappedM = new MWrappedM(network, dexKey, dexHelper);
+      if (mWrappedM.initializePricing) {
+        await mWrappedM.initializePricing(blockNumber);
+      }
+    });
+
+    it('getPoolIdentifiers and getPricesVolume SELL', async function () {
+      await testPricingOnNetwork(
+        mWrappedM,
+        network,
+        dexKey,
+        blockNumber,
+        'M',
+        'WrappedM',
+        SwapSide.SELL,
+        amountsForSell,
+        '',
+      );
+    });
+
+    it('getPoolIdentifiers and getPricesVolume BUY', async function () {
+      await testPricingOnNetwork(
+        mWrappedM,
+        network,
+        dexKey,
+        blockNumber,
+        'M',
+        'WrappedM',
+        SwapSide.BUY,
+        amountsForSell,
+        '',
+      );
+    });
+
+    it('getTopPoolsForToken: M', async function () {
+      const tokenA = Tokens[network]['M'];
+      const dexHelper = new DummyDexHelper(network);
+      const mWrappedM = new MWrappedM(network, dexKey, dexHelper);
+
+      const poolLiquidity = await mWrappedM.getTopPoolsForToken(
+        tokenA.address,
+        10,
+      );
+      console.log(
+        `${tokenA.symbol} Top Pools:`,
+        JSON.stringify(poolLiquidity, null, 2),
+      );
+
+      checkPoolsLiquidity(poolLiquidity, tokenA.address, dexKey);
+    });
+
+    it('getTopPoolsForToken: WrappedM', async function () {
+      const tokenA = Tokens[network]['WrappedM'];
+      const dexHelper = new DummyDexHelper(network);
+      const mWrappedM = new MWrappedM(network, dexKey, dexHelper);
+
+      const poolLiquidity = await mWrappedM.getTopPoolsForToken(
         tokenA.address,
         10,
       );
